+++ dataflat/polars/__init__.py
@@ -1,0 +1,3 @@
+from dataflat.polars.flattener import CustomFlattener
+
+__all__ = ["CustomFlattener"]