+++ dataflat/pyarrow/__init__.py
@@ -1,0 +1,3 @@
+from dataflat.pyarrow.flattener import CustomFlattener
+
+__all__ = ["CustomFlattener"]